WIP: working on test fixes

+++ models/foo_models.py
@@ -3,11 +3,13 @@ from utils.utils import get_utc_now
 from database import Base
 from datetime import datetime
 
+from exceptions import MyFastAPIAppException
+
 
 class Foo(Base):
     __tablename__ = "foo"
     id = Column(BigInteger, primary_key=True, autoincrement=True)
-    name = Column(String, nullable=False)
+    name = Column(String, nullable=False, unique=True)
     timestamp = Column(DateTime, nullable=False, default=get_utc_now)
 
     def __init__(self, name: str, timestamp: datetime = None):
@@ -19,7 +21,7 @@ class Foo(Base):
 
     @classmethod
     def get(cls, session, id: int):
-        return session.query(cls).filter(cls.id == id).one()
+        return session.query(cls).filter(cls.id == id).one_or_none()
 
     @classmethod
     def get_all(cls, session):
@@ -34,6 +36,9 @@ class Foo(Base):
 
     @classmethod
     def delete(cls, session, id: int):
+        existing = cls.get(session, id=id)
+        if not existing:
+            raise MyFastAPIAppException(f"Foo with id {id} not found")
         session.query(cls).filter(cls.id == id).delete()
         session.flush()
         return True
